MAIN/requirements.txt
pyttsx3
groq
pdfminer.six
python-dotenv
fastapi
uvicorn[standard]
python-multipart 